// src/interfaces/styles/messageCustomStyle.ts
export interface MessageCustomStyle {
  message_bg_color_sender: string;
  message_color_text_sender: string;
  message_bg_color_received: string;
  message_color_text_received: string;
  message_text_color_datetime: string;
  message_bg_color_loading: string;
  message_color_text_loading: string;
}